Update LibItemSearch-1.2
The latest change to this file was adding the Lib.Filters to check for "Usable". It was named Lib.Filters.quality, identical to filter used to search for quality. This entirely breaks the possibility to search for items by their quality.

This could be adressed in the function that adds these filters to allow several filters with identical name.

A simpler change is to change the filter name that checks for usable, to Lib.Filters.usable. 

This allows searching for both usable and quality.